Finished refining topic_modelling
